修正数据库和模板文件
把数据库主键id统一换成long,模板后缀改为html

+++ src/main/java/xyz/oilpea/schooldemo/demo/dao/AlumniMapper.java
@@ -3,13 +3,13 @@ package xyz.oilpea.schooldemo.demo.dao;
 import xyz.oilpea.schooldemo.demo.pojo.Alumni;
 
 public interface AlumniMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Alumni record);
 
     int insertSelective(Alumni record);
 
-    Alumni selectByPrimaryKey(Integer id);
+    Alumni selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Alumni record);
